# Optimize Random{NumberGenerator}.GetItems for power-of-two choices