--- a/src/libraries/System.Private.CoreLib/src/System/Random.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Random.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Diagnostics;
+using System.Numerics;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace System
 {
@@ -194,6 +196,38 @@ namespace System
                 throw new ArgumentException(SR.Arg_EmptySpan, nameof(choices));
             }
 
+            // The most expensive part of this operation is the call to get random data. We can
+            // do so only once rather than per element in a common case:
+            // - The number of choices is <= 256. This let's us get a single byte per choice.
+            // - The number of choices is a power of two. This let's us a byte and simply mask off
+            //   unnecessary bytes cheaply rather than needing to use rejection sampling.
+            // - The type of T doesn't contain any references and is larger than a byte in size.
+            //   This let's us write random bytes into the destination buffer rather than needing
+            //   a scratch buffer. we can use a single RNG.Fill to get all of the required randomness,
+            //   and then walk backwards through the destination, filling in the corresponding choice
+            //   based on the random byte at that position when interpreted as bytes. By walking backwards,
+            //   we won't overrite any random bytes that we still need to read.
+            if (!RuntimeHelpers.IsReferenceOrContainsReferences<T>() && Unsafe.SizeOf<T>() >= 2 &&
+                BitOperations.IsPow2(choices.Length) && choices.Length <= 256)
+            {
+                // Reinterpret the bottom destination.Length bytes of the destination as our
+                // scratch buffer for randomness.
+                Span<byte> randomBytesInDestination = new Span<byte>(
+                    ref Unsafe.As<T, byte>(ref MemoryMarshal.GetReference(destination)),
+                    destination.Length);
+                NextBytes(randomBytesInDestination);
+
+                int mask = choices.Length - 1;
+                for (int i = destination.Length - 1; i >= 0; i--)
+                {
+                    destination[i] = choices[randomBytesInDestination[i] & mask];
+                }
+
+                return;
+            }
+
+            // Simple fallback: get each item individually, generating a new random Int32 for each
+            // item. This is slower than the above, but it works for all types and sizes of choices.
             for (int i = 0; i < destination.Length; i++)
             {
                 destination[i] = choices[Next(choices.Length)];