--- a/src/libraries/System.Private.CoreLib/src/System/Random.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Random.cs
@@ -201,14 +201,14 @@ namespace System
             // - The number of choices is <= 256. This let's us get a single byte per choice.
             // - The number of choices is a power of two. This let's us a byte and simply mask off
             //   unnecessary bytes cheaply rather than needing to use rejection sampling.
-            // - The type of T doesn't contain any references and is larger than a byte in size.
-            //   This let's us write random bytes into the destination buffer rather than needing
-            //   a scratch buffer. we can use a single RNG.Fill to get all of the required randomness,
-            //   and then walk backwards through the destination, filling in the corresponding choice
-            //   based on the random byte at that position when interpreted as bytes. By walking backwards,
-            //   we won't overrite any random bytes that we still need to read.
-            if (!RuntimeHelpers.IsReferenceOrContainsReferences<T>() && Unsafe.SizeOf<T>() >= 2 &&
-                BitOperations.IsPow2(choices.Length) && choices.Length <= 256)
+            // - The type of T doesn't contain any references. This let's us write random bytes into the
+            //   destination buffer rather than needing a scratch buffer. we can use a single RNG.Fill to get
+            //   all of the required randomness, and then walk backwards through the destination, filling in the
+            //   corresponding choice based on the random byte at that position when interpreted as bytes. By
+            //   walking backwards, we won't overrite any random bytes that we still need to read.
+            if (!RuntimeHelpers.IsReferenceOrContainsReferences<T>() &&
+                BitOperations.IsPow2(choices.Length) &&
+                choices.Length <= 256)
             {
                 // Reinterpret the bottom destination.Length bytes of the destination as our
                 // scratch buffer for randomness.