--- a/src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/RandomNumberGenerator.cs
+++ b/src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/RandomNumberGenerator.cs
@@ -2,11 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Buffers;
-using System.Buffers.Binary;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Numerics;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 namespace System.Security.Cryptography
@@ -354,36 +352,30 @@ namespace System.Security.Cryptography
         private static void GetItemsCore<T>(ReadOnlySpan<T> choices, Span<T> destination)
         {
             // The most expensive part of this operation is the call to get random data. We can
-            // do so only once rather than per element in a common case:
-            // - The number of choices is <= 256. This let's us get a single byte per choice.
-            // - The number of choices is a power of two. This let's us use a byte and simply mask off
+            // do so potentially many fewer times if:
+            // - the number of choices is <= 256. This let's us get a single byte per choice.
+            // - the number of choices is a power of two. This let's us use a byte and simply mask off
             //   unnecessary bits cheaply rather than needing to use rejection sampling.
+            // In such a case, we can grab a bunch of random bytes in one call.
             if (BitOperations.IsPow2(choices.Length) && choices.Length <= 256)
             {
-                const int StackAllocThreshold = 256;
-
-                // Get a scratch buffer to fill with random bytes, either stack space if the
-                // destination is small enough, or an array pool array if it's larger.
-                // If T is unmanaged, we _could_ use the destination buffer itself, reinterpreted
-                // as a byte buffer, to hold the scratch bytes, but then we'd potentially be temporarily
-                // writing invalid Ts into the destination, which is a risk.
-                byte[]? arrayPoolArray = null;
-                Span<byte> randomBytes = destination.Length <= StackAllocThreshold ?
-                    stackalloc byte[StackAllocThreshold] :
-                    (arrayPoolArray = ArrayPool<byte>.Shared.Rent(destination.Length));
-                randomBytes = randomBytes.Slice(0, destination.Length);
-
-                Fill(randomBytes);
-
-                int mask = choices.Length - 1;
-                for (int i = 0; i < destination.Length; i++)
+                Span<byte> randomBytes = stackalloc byte[512]; // arbitrary size, a balance between stack consumed and number of random calls required
+                while (!destination.IsEmpty)
                 {
-                    destination[i] = choices[randomBytes[i] & mask];
-                }
+                    if (destination.Length < randomBytes.Length)
+                    {
+                        randomBytes = randomBytes.Slice(0, destination.Length);
+                    }
 
-                if (arrayPoolArray is not null)
-                {
-                    ArrayPool<byte>.Shared.Return(arrayPoolArray);
+                    RandomNumberGeneratorImplementation.FillSpan(randomBytes);
+
+                    int mask = choices.Length - 1;
+                    for (int i = 0; i < randomBytes.Length; i++)
+                    {
+                        destination[i] = choices[randomBytes[i] & mask];
+                    }
+
+                    destination = destination.Slice(randomBytes.Length);
                 }
 
                 return;