When possible, make a single call to get the required randomness rather than one per element.

Closes https://github.com/dotnet/runtime/issues/92130

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Security.Cryptography;

BenchmarkSwitcher.FromAssembly(typeof(Tests).Assembly).Run(args);

[HideColumns("Error", "StdDev", "Median", "RatioSD")]
public class Tests
{
    private const int Length = 11;
    private static readonly string s_charChoices = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_-";
    private static ReadOnlySpan<int> s_intChoices => [42, 43, 44, 45, 46, 47, 48, 49];
    private static int[] s_dest = new int[Length];

    [Benchmark]
    public string RngGetString() => RandomNumberGenerator.GetString(s_charChoices, Length);

    [Benchmark]
    public void RandGetItems() => Random.Shared.GetItems(s_intChoices, s_dest);
}
```

| Method       | Toolchain         | Mean      | Ratio |
|------------- |------------------ |----------:|------:|
| RngGetString | \main\corerun.exe | 764.18 ns |  1.00 |
| RngGetString | \pr\corerun.exe   |  98.25 ns |  0.13 |
|              |                   |           |       |
| RandGetItems | \main\corerun.exe |  57.21 ns |  1.00 |
| RandGetItems | \pr\corerun.exe   |  16.56 ns |  0.29 |