@Ponant 2023-09-18
I few remarks if you do not mind.

1) I tested @bartonjs method against mine and confirm it is faster, but it is mainly due to the use bitwise AND instead of integer division, and not theunsafe territory. At least on my machine.

2) I think the benefit of the pow2 non-negligible speedup should be documented in th emethods that use it such as GetString, perhaps...

3) I have an addional suggestion, which I think can be beneficial to users and which can be implemented in the same PR. But I do not know if oyu prefer that I create another issue or drop it here? it is about rewritting GetBytes to allow users to have a byte arrays where each byte is bound to [0,N] where N is a power of two. For N=256, we recover the case of the current getBytes. Let me know I will drop a code