@Ponant 2023-09-18
> > I have an addional suggestion, which I think can be beneficial to users and which can be implemented in the same PR. But I do not know if oyu prefer that I create another issue or drop it here? it is about rewritting GetBytes to allow users to have a byte arrays where each byte is bound to [0,N] where N is a power of two. For N=256, we recover the case of the current getBytes. Let me know I will drop a code
> 
> This sounds very niche, and the savings is effectively an array access per element. I don't think it's worth special-casing without seeing more examples of where it would be really meaningful.
> 

A) Generate random booleans
B) Generate Random vectors in 4 dimensions (my case)