@GrabYourPitchforks 2023-09-18
@stephentoub Consider also duplicating the test here with one that passes a power-of-two sized values span. That way we get distribution tests across both implementations which now exist in this method.

https://github.com/dotnet/runtime/blob/0ad4e699b0a6d92c2d08f7b9b6393448ab8f310a/src/libraries/System.Security.Cryptography/tests/RandomNumberGeneratorTests.cs#L669-L675