@Ponant 2023-09-18
The current `GetBytes` returns and array of bytes, where each byte can take any value from 0 to 255. My proposal is to have GetBytes to also allow me to return an array of bytes but where each byte is bound to [0,N] where N is a powere of two. For example, I can generate a byte array of bools  011100011010101 or random quaternions or any 4d vector, or octonions or above powers.

I am using something like that, it is essentially the same spirit as you do here with the new GetString in this PR, but for getting byte arrays. I am using something like that (checks on parameters omitted for brievety)

```
private static byte[] GetRandomBytes(int length, int maxNumOfBytes)
 {
        Span<byte> bytes = length <= _stackAllocThreshold ? stackalloc byte[length] : new byte[length];

        // Generate random bytes, each taking values between 0 and 255
        RandomNumberGenerator.Fill(bytes);

        // We do not need to loop if we cover all byte possibilities, since  RandomNumberGenerator.Fill does this 
        if (maxNumOfBytes != 256)
        {
            byte mask = (byte)(maxNumOfBytes - 1);

            // make each byte between 0 and maxNumOfBytes
            for (var i = 0; i < bytes.Length; i++)
                bytes[i] &= mask;

        }

        return bytes.ToArray();
}
```