@stephentoub 2023-09-18
> The current GetBytes returns an array of bytes, where each byte can take any value from 0 to 255. My proposal is to have GetBytes to also allow me to return an array of bytes but where each byte is bound to [0,N] where N is a power of two.

That's a new API.  Feel free to open a separate proposal for one.  That won't be done as part of an optimization PR.