@vcsjones 2023-09-18
> The current GetBytes returns an array of bytes, where each byte can take any value from 0 to 255. My proposal is to have GetBytes to also allow me to return an array of bytes but where each byte is bound to [0,N] where N is a power of two. For example, I can generate a byte array of bools 011100011010101 or random quaternions or any 4d vector, or octonions or above powers.

This would be a new [API proposal](https://github.com/dotnet/runtime/blob/main/docs/project/api-review-process.md).

We can discuss the merits and usefulness of the proposal there to keep this PR focused on the perf of existing APIs.