@stephentoub 2024-09-19
@bartonjs, per your feedback on my recent PR, I've marked this as a breaking change to be documented.