@stephentoub 2024-09-19
Opened https://github.com/dotnet/runtime/pull/108018 to revert the breaking change aspect.